ui/requirements.txt
mysql-connector-python==8.1.0
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
pika==1.3.2
